requirements.txt
pyserial>=3.4
pyserial-asyncio>=0.4

